requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.1
